+++ src/span.cpp
@@ -805,7 +805,7 @@ namespace pinpoint {
 
         data_->setErrorFuncId(agent_->cacheError(error_name));
         data_->setErrorString(abbreviateErrorString(error_message));
-        markSpanError(error_name, error_message);
+        markSpanError(ErrorCategory::kException, error_name, error_message);
     } CATCH_AND_LOG("set error")
 
     void SpanImpl::SetStatusCode(int status) {
@@ -813,7 +813,7 @@ namespace pinpoint {
 
         data_->getAnnotations()->AppendInt(ANNOTATION_HTTP_STATUS_CODE, status);
         if (isStatusFail(status)) {
-            markSpanError();
+            markSpanError(ErrorCategory::kHttpStatus);
         }
     }